// src/component-library-contentsdk/site-components/QuoteBlockCenter/QuoteBlockCenter.types.ts
import { ReactNode, type JSX } from 'react';

export interface QuoteBlockCenterProps {
  children?: ReactNode | JSX.Element | string;
  author?: {
    name?: ReactNode | JSX.Element | string;
    image?: ReactNode | JSX.Element;
    tag?: ReactNode | JSX.Element | string;
  };
  alignment?: 'left' | 'center';
}